+++ resource-server/src/main/java/com/elearning/common/TimeUtils.java
@@ -1,0 +1,12 @@
+package com.elearning.common;
+
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+public interface TimeUtils {
+
+    static final DateTimeFormatter FORMATTER = DateTimeFormatter
+            .ofPattern("yyyy-MM-dd HH:mm:ss")
+            .withZone(ZoneId.systemDefault());
+
+}
