--- resource-server/src/main/java/com/elearning/common/config/JacksonCustomizations.java
@@ -1,72 +1,0 @@
-package com.elearning.common.config;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.Module;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.deser.ValueInstantiator;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import org.javamoney.moneta.Money;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.i18n.LocaleContextHolder;
-
-import javax.money.MonetaryAmount;
-import javax.money.format.MonetaryFormats;
-import java.io.IOException;
-
-@Configuration(proxyBeanMethods = false)
-public class JacksonCustomizations {
-
-    @Bean
-    Module moneyModule() {
-        return new MoneyModule();
-    }
-
-    static class MoneyModule extends SimpleModule {
-
-        public MoneyModule() {
-            addSerializer(MonetaryAmount.class, new MonetaryAmountSerializer());
-            addValueInstantiator(MonetaryAmount.class, new MoneyInstantiator());
-        }
-
-        static class MonetaryAmountSerializer extends StdSerializer<MonetaryAmount> {
-
-            public MonetaryAmountSerializer() {
-                super(MonetaryAmount.class);
-            }
-
-            @Override
-            public void serialize(MonetaryAmount value, JsonGenerator gen, SerializerProvider provider)
-                    throws IOException {
-                if (value != null) {
-                    gen.writeString(MonetaryFormats.getAmountFormat(LocaleContextHolder.getLocale()).format(value));
-                } else {
-                    gen.writeNull();
-                }
-            }
-
-        }
-
-        static class MoneyInstantiator extends ValueInstantiator {
-
-            @Override
-            public String getValueTypeDesc() {
-                return MonetaryAmount.class.toString();
-            }
-
-            @Override
-            public boolean canCreateFromString() {
-                return true;
-            }
-
-            @Override
-            public Object createFromString(DeserializationContext context, String value) throws IOException {
-                return Money.parse(value, MonetaryFormats.getAmountFormat(LocaleContextHolder.getLocale()));
-            }
-        }
-
-    }
-
-}
