+++ resource-server/src/main/java/com/elearning/common/config/jackson/JacksonCustomizations.java
@@ -1,0 +1,20 @@
+package com.elearning.common.config.jackson;
+
+import com.fasterxml.jackson.databind.Module;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration(proxyBeanMethods = false)
+public class JacksonCustomizations {
+
+    @Bean
+    Module moneyModule() {
+        return new MoneyModule();
+    }
+
+    @Bean
+    Module dateTimeModule() {
+        return new DateTimeModule();
+    }
+
+}
